[FLINK-21030][runtime] Adds trigger for global failover
We cannot assume that the termination future is finishing when triggering a
synchronous savepoint. It might be that the savepoint creation succeeds but
the graceful shutdown of the job fails. In this case, the job termination does
not complete. The job might end up in an inconsistent state. That's why, a
global fail over is triggered if the scheduler observes one of the current
executions ending up in a non-finished state.

Unit tests and integration tests are added to cover this use-case.